refactor: group all views under one directory and separate
