+++ authentication/views/__init__.py
@@ -1,0 +1,7 @@
+from .password import SignUpEndpoint, LoginEndpoint
+
+from .verification import VerifyEmailEndpoint, ResendVerificationEndpoint
+
+from .password_management import ResetPasswordEndpoint
+
+from .logout import LogoutEndpoint
