+++ authentication/views/logout.py
@@ -1,0 +1,30 @@
+from rest_framework import status
+from rest_framework.exceptions import ValidationError
+from rest_framework.generics import GenericAPIView
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+from rest_framework_simplejwt.tokens import RefreshToken, TokenError
+
+from authentication.serializers import LogoutSerializer
+from authentication.services import get_user_by_refresh_token, update_user_logout_metadata
+
+
+class LogoutEndpoint(GenericAPIView):
+    """Logout endpoint for all auth methods"""
+    serializer_class = LogoutSerializer
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request):
+        refresh_token = request.data.get("refresh")
+
+        try:
+            # Retrieve the owner
+            user = get_user_by_refresh_token(refresh_token)
+            token = RefreshToken(refresh_token)
+            token.blacklist()
+        except TokenError:
+            raise ValidationError({"error": "Token is invalid or already blacklisted."})
+
+        updated_user = update_user_logout_metadata(user, request)
+
+        return Response(status=status.HTTP_204_NO_CONTENT)
